requirements.txt
streamlit==1.28.1
googlemaps==4.10.0
folium==0.14.0
streamlit-folium==0.15.0
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.0